+++ src/core/util/automation/automaton.rs
@@ -755,9 +755,9 @@ impl<'a> InPlaceMergeSorterImpl<'a> {
     }
 }
 impl Sorter for InPlaceMergeSorterImpl<'_> {
-    fn compare(&mut self, i: i32, j: i32) -> Result<i32> {
-        let i_start = i as usize * 4;
-        let j_start = j as usize * 4;
+    fn compare(&mut self, i: usize, j: usize) -> Result<i32> {
+        let i_start = i * 4;
+        let j_start = j * 4;
 
         // First src
         let i_src = self.transitions[i_start];
@@ -822,9 +822,9 @@ impl<'a> MinMaxDestSorter<'a> {
     }
 }
 impl Sorter for MinMaxDestSorter<'_> {
-    fn compare(&mut self, i: i32, j: i32) -> Result<i32> {
-        let i_start = 3 * i as usize;
-        let j_start = 3 * j as usize;
+    fn compare(&mut self, i: usize, j: usize) -> Result<i32> {
+        let i_start = 3 * i;
+        let j_start = 3 * j;
 
         // First compare min
         let i_min = self.transitions[i_start + 1];
@@ -876,9 +876,9 @@ impl<'a> DestMinMaxSorter<'a> {
     }
 }
 impl Sorter for DestMinMaxSorter<'_> {
-    fn compare(&mut self, i: i32, j: i32) -> Result<i32> {
-        let i_start = (3 * i) as usize;
-        let j_start = (3 * j) as usize;
+    fn compare(&mut self, i: usize, j: usize) -> Result<i32> {
+        let i_start = 3 * i;
+        let j_start = 3 * j;
 
         // First dest:
         let i_dest = self.transitions[i_start];
